backend/lambda/yolo_analysis/requirements.txt
# Lambda Layer 또는 배포 패키지에 포함
ultralytics>=8.0.0
Pillow>=10.0.0
numpy>=1.24.0
# boto3, urllib are provided by Lambda runtime
